user/requirements.txt
werkzeug
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
pymysql
PyJWT==1.4.2
